feat(data): finish trial metadata's four surfaces, add saveable design presets, and bring edit-dataset to parity with add-dataset
Trial metadata (DATA-29): reaches the CLI (scanpath-studio render
--trial-metadata / --trial-metadata-reader-column), the headless API
(api.load_trial_metadata), the export bundle (metadata/trials.*), the
chip strip, and the add-dataset wizard (previously edit-dataset only).
A metadata table that joins to nothing now raises instead of reporting
"Joined to 0" as a quiet success (DATA-20, DATA-29), a chip whose
recorded column is blank for a trial now falls through to the next
source instead of stopping there, and detaching the trial table now
actually detaches it instead of re-attaching itself from a stale
uploader key.

Design presets (VIZ-39): plot settings can be saved, named, and
restored from a library ("My designs") instead of the single unnamed
"Custom" slot quick views offered before, following the on-device
recovery cache and Save & restore the same way annotations do.

Edit-dataset/add-dataset parity (UX-104, UX-106, UX-107): the editor
now draws the same per-table field blocks as the add screen, offers an
uploader to fill in an AOI or fixations table a dataset was added
without, gates text-size and character-AOI options the same way the
add screen does, and matches its button placement/behavior (Save
returns to the dataset list, Cancel only confirms when something
actually changed).

Also: metadata joins report one summary line with a "Join details"
expander instead of four paragraphs (UX-105); plot-rail section titles
and their popover triggers no longer show redundant hover text
(UX-103) or grow a spurious scrollbar (UX-102); add-dataset's mapping
pickers now offer every column in the uploaded file rather than only
the ones a previous upload for the same filename happened to keep
(UX-108); docs updated to match the app as of 0.29.0.

See CHANGELOG.md for the full per-item write-up.

+++ scanpath_studio/export_status.py
@@ -72,7 +72,7 @@ def format_duration(seconds: float | None) -> str:
         return ""
     if seconds < 1:
         return "under a second"
-    seconds = int(round(seconds))
+    seconds = round(seconds)  # round() on a float already returns an int
     hours, rest = divmod(seconds, 3600)
     minutes, secs = divmod(rest, 60)
     if hours:
